refactor: Result 패키지 변경
