+++ src/main/java/blackjack/domain/participant/DealerResult.java
@@ -1,5 +1,7 @@
 package blackjack.domain.participant;
 
+import blackjack.domain.Result;
+
 import java.util.Collections;
 import java.util.EnumMap;
 import java.util.Map;
