+++ src/test/java/blackjack/domain/ProfitTest.java
@@ -1,6 +1,5 @@
 package blackjack.domain;
 
-import blackjack.domain.participant.Result;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
